10.43.101.182/requirements.txt
fastapi
uvicorn
mlflow
pydantic
python-dotenv
pandas
gradio
boto3